+++ src/handlers/events/delivery/verify/executeGoal.ts
@@ -15,6 +15,7 @@ export async function executeGoal(execute: GoalExecutor,
     await markGoalInProcess(ctx, sdmGoal, params.goal);
     try {
         const result = await execute(status, ctx, params);
+        logger.info("Result of %s: %j", params.implementationName, result);
         await markStatus(ctx, sdmGoal, params.goal, result);
         return Success;
     } catch (err) {
